+++ src/Phan/Analysis/AssignOperatorFlagVisitor.php
@@ -269,17 +269,10 @@ class AssignOperatorFlagVisitor extends FlagVisitorImplementation
             $probably_array_type = UnionType::of([ArrayType::instance(false)], $int_or_float_or_array);
             $unknown_type = UnionType::of([], $int_or_float_or_array);
         }
-        $left = UnionTypeVisitor::unionTypeFromNode(
-            $this->code_base,
-            $this->context,
-            $node->children['var']
-        );
-
-        $right = UnionTypeVisitor::unionTypeFromNode(
-            $this->code_base,
-            $this->context,
-            $node->children['expr']
-        );
+        $code_base = $this->code_base;
+        $context = $this->context;
+        $left = UnionTypeVisitor::unionTypeFromNode($code_base, $context, $node->children['var']);
+        $right = UnionTypeVisitor::unionTypeFromNode($code_base, $context, $node->children['expr']);
 
         // fast-track common cases
         if ($left->isNonNullIntType()
@@ -308,24 +301,24 @@ class AssignOperatorFlagVisitor extends FlagVisitorImplementation
         }
 
         $left_is_array = (
-            !$left->genericArrayElementTypes()->isEmpty()
+            !$left->genericArrayElementTypes(false, $code_base)->isEmpty()
             && $left->nonArrayTypes()->isEmpty()
         ) || $left->isType(ArrayType::instance(false));
 
         $right_is_array = (
-            !$right->genericArrayElementTypes()->isEmpty()
+            !$right->genericArrayElementTypes(false, $code_base)->isEmpty()
             && $right->nonArrayTypes()->isEmpty()
         ) || $right->isType(ArrayType::instance(false));
 
         if ($left_is_array
             && !$right->canCastToUnionType(
                 ArrayType::instance(false)->asPHPDocUnionType(),
-                $this->code_base
+                $code_base
             )
         ) {
             Issue::maybeEmit(
-                $this->code_base,
-                $this->context,
+                $code_base,
+                $context,
                 Issue::TypeInvalidRightOperand,
                 $node->lineno ?? 0
             );
@@ -333,12 +326,12 @@ class AssignOperatorFlagVisitor extends FlagVisitorImplementation
         } elseif ($right_is_array
             && !$left->canCastToUnionType(
                 ArrayType::instance(false)->asPHPDocUnionType(),
-                $this->code_base
+                $code_base
             )
         ) {
             Issue::maybeEmit(
-                $this->code_base,
-                $this->context,
+                $code_base,
+                $context,
                 Issue::TypeInvalidLeftOperand,
                 $node->lineno ?? 0
             );